# server/swagger.yaml
openapi: 3.0.0
info:
  title: Application de covoiturage universitaire
  version: 1.0.0
  description: API pour publier et rechercher des itinéraires universitaires

servers:
  - url: http://localhost:8000/api

tags:
  - name: Users
  - name: Itinaries
  - name: Bookings
  - name: Destinations

paths:
  /user/register:
    post:
      tags:
        - Users
      summary: S'inscrire sur le site
      requestBody:
        description: Informations d'inscription
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Utilisateur enregistré avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Création de l'utilisateur impossible

  /user/login:
    post:
      tags:
        - Users
      summary: Récupérer tout les utilisateurs et leurs informations
      requestBody:
        description: Informations d'inscription
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Récuperation de la listes des utilisateur enregistré avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLogin"
        "404":
          description: Création de l'utilisateur impossible

  /user/logout:
    post:
      tags:
        - Users
      summary: Déconnecter l' utilisateur
      requestBody:
        description: Déconnecter l'utilisateur
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Confirmation de la déconnexion de l'utilisateur avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLogin"
        "404":
          description: Déconnexion de l'utilisateur impossible

  /users:
    get:
      tags:
        - Users
      summary: Obtenir la liste des utilisateurs et leurs informations
      responses:
        "200":
          description: Liste des utilisateurs récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          description: Impossible de récupérer la liste des utilisateurs disponibles

  /itinaries:
    post:
      tags:
        - Itinaries
      summary: Créer un itinéraire
      requestBody:
        description: Informations de l'itinéraire
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Itinaries"
      responses:
        "200":
          description: Itinéraire créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de créer l'itinéraire

    get:
      tags:
        - Itinaries
      summary: Obtenir la liste des itinéraires disponibles
      responses:
        "200":
          description: Liste des itinéraires récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de récupérer la liste des itinéraires disponibles

  /itinariesFormatted/:
    get:
      tags:
        - Itinaries
      summary: Obtenir la liste des itinéraires disponibles formattés
      responses:
        "200":
          description: Liste des itinéraires récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de récupérer la liste des itinéraires disponibles

  /itinaries/startAddress/{startAddress}:
    get:
      tags:
        - Itinaries
      summary: Obtenir la liste des itinéraires disponibles pour une adresse de départ
      parameters:
        - name: startAddress
          in: path
          description: Addresse de départ
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liste des itinéraires récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de récupérer la liste des itinéraires disponibles

  /itinaries/email/{email}:
    get:
      tags:
        - Itinaries
      summary: Obtenir la liste des itinéraires disponibles pour un user
      parameters:
        - name: email
          in: path
          description: le user
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liste des itinéraires récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de récupérer la liste des itinéraires disponibles

  /itinaries/destination/{destination}:
    get:
      tags:
        - Itinaries
      summary: Obtenir les informations de la destination universitaire disponible
      parameters:
        - name: destination
          in: path
          description: Destination
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liste des destinations universitaires
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Destinations introuvable

  /itinaries/PassengerList/{itinaries_id}:
    get:
      tags:
        - Itinaries
      summary: Obtenir les informations de la destination universitaire disponible
      parameters:
        - name: itinaries_id
          in: path
          description: itinaries_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liste des destinations universitaires
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Destinations introuvable

  /itinarie/{itinaries_id}:
    put:
      tags:
        - Itinaries
      summary: Mettre à jour un itinéraire existant
      parameters:
        - in: path
          name: itinaries_id
          required: true
          description: ID de l'itinéraire à mettre à jour
          schema:
            type: integer
      requestBody:
        description: Informations de l'itinéraire mises à jour
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Itinaries"
      responses:
        "200":
          description: Itinéraire mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de mettre à jour l'itinéraire

    delete:
      tags:
        - Itinaries
      summary: Supprimer un itinéraire existant
      parameters:
        - in: path
          name: itinaries_id
          required: true
          description: ID de l'itinéraire à supprimer
          schema:
            type: integer
      responses:
        "200":
          description: Itinéraire supprimé avec succès
        "404":
          description: Impossible de supprimer l'itinéraire

  /itinarie/{itinaries_id}/seatsmin:
    patch:
      tags:
        - Itinaries
      summary: Mettre à jour le nombre de places disponibles dans un itinéraire existant pour un conducteur
      parameters:
        - name: itinaries_id
          in: path
          description: ID de l'itinéraire
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: nombre de places mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de mettre à jour le nombre de places

  /itinarie/{itinaries_id}/seatsplus:
    patch:
      tags:
        - Itinaries
      summary: Mettre à jour le nombre de places disponibles dans un itinéraire existant pour un conducteur
      parameters:
        - name: itinaries_id
          in: path
          description: ID de l'itinéraire
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: nombre de places mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries"
        "404":
          description: Impossible de mettre à jour le nombre de places

  /bookings:
    post:
      tags:
        - Bookings
      summary: Réserver une place dans un itinéraire existant
      requestBody:
        description: Informations de réservation
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Itinaries_User"
      responses:
        "200":
          description: Réservation effectuée avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries_User"
        "404":
          description: Réservation impossible

    get:
      tags:
        - Bookings
      summary: Obtenir toutes les réservations effectuées par tout les utilisateur
      responses:
        "200":
          description: Réservations récupérées avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinaries_User"

  /bookings/userConnected:
    get:
      tags:
        - Bookings
      summary: Obtenir toutes les informations des réservation du user connecté
      responses:
        "200":
          description: Liste des destinations universitaires
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries_User"
        "404":
          description: Destinations introuvable

  /booking/{itinaries_user_id}:
    get:
      tags:
        - Bookings
      summary: Obtenir toutes les informations d'une réservation
      parameters:
        - name: itinaries_user_id
          in: path
          description: Id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liste des destinations universitaires
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries_User"
        "404":
          description: Destinations introuvable
    delete:
      tags:
        - Bookings
      summary: Supprimer une réservation existante
      parameters:
        - in: path
          name: itinaries_user_id
          required: true
          description: ID de la réservation à supprimer
          schema:
            type: integer
      responses:
        "200":
          description: Itinéraire supprimé avec succès
        "404":
          description: Impossible de supprimer l'itinéraire

  /booking/user/{fk_user}/itinarie/{fk_itinaries}:
    delete:
      tags:
        - Bookings
      summary: Supprimer une réservation existante
      parameters:
        - in: path
          name: fk_user
          required: true
          description: ID de la réservation à supprimer
          schema:
            type: string
        - in: path
          name: fk_itinaries
          required: true
          description: ID de la réservation à supprimer
          schema:
            type: string

      responses:
        "200":
          description: Itinéraire supprimé avec succès
        "404":
          description: Impossible de supprimer l'itinéraire

  /booking/{itinaries_user_id}/accept:
    patch:
      tags:
        - Bookings
      summary: Accepter un passager
      parameters:
        - name: itinaries_user_id
          in: path
          description: ID de la réservation
          required: true
          schema:
            type: string
      # requestBody:
      #   description: Informations de l'itinéraire mises à jour
      #   required: true
      #   content:
      #     application/json:
      #       schema:
      #         type: object
      #         properties:
      #           request_user:
      #             type: boolean
      responses:
        "200":
          description: réservation mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries_User"
        "404":
          description: Impossible de mettre à jour la réservation

  /booking/{itinaries_user_id}/deny:
    patch:
      tags:
        - Bookings
      summary: Refuser un passager
      parameters:
        - name: itinaries_user_id
          in: path
          description: ID de la réservation
          required: true
          schema:
            type: string
      # requestBody:
      #   description: Modifier le status d'un passager
      #   required: true
      #   content:
      #     application/json:
      #       schema:
      #         type: object
      #         properties:
      #           request_user:
      #             type: boolean
      responses:
        "200":
          description: réservation mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinaries_User"
        "404":
          description: Impossible de mettre à jour la réservation

  # /destinations:
  #   post:
  #     tags:
  #       - Destinations
  #     summary: Créer une nouvelle destination
  #     requestBody:
  #       description: Informations de la destination
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/Destination"
  #     responses:
  #       "200":
  #         description: Creation de destiantion effectuée avec succès
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Destination"
  #       "404":
  #         description: Creation de destination impossible
  #   get:
  #     tags:
  #       - Destinations
  #     summary: Obtenir toutes les destinations universitaires disponibles
  #     responses:
  #       "200":
  #         description: Liste de toutes les destinations universitaires
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Destination"
  #       "404":
  #         description: Destinations introuvable

components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        address:
          type: string
        number_passengers_max:
          type: integer
        lisence_plate:
          type: string
        picture:
          type: string

    Itinaries:
      type: object
      properties:
        startAddress:
          type: string
        seats:
          type: integer
        destination:
          type: string
        startDate:
          type: string
        hours:
          type: string

    Itinaries_User:
      type: object
      properties:
        fk_itinaries:
          type: integer
        fk_user:
          type: string
        type_user:
          type: string
        request_user:
          type: boolean
        message:
          type: string

    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string